First version of the articles system
Articles can be published via creating a new post in the post folder. For the moment, only the 6 most recent articles are displayed on the homepage; There is no reference to existing article except the one in the homepage. It would be nice to build something like an index in the future; For now, articles only refer to the homepage. They might be published with ou without a cover picture. Also, a small synthesis is expected and would be displayed in the homepage.